+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
 ## version 0.0.2
 
-* add mkCookie function for convenient cookie construction
+* added mkCookie function for convenient cookie construction
+* added the focus function which changes focus of windows and frames; focusWindow was removed as focus is a generalization of it